carritoPrecio

+++ stock_manager/src/main/java/es/deusto/spq/data/Cesta.java
@@ -121,7 +121,7 @@ public class Cesta implements Serializable {
    //A modificar por Oferta
    public float getRecibo() {
 
-       int r = 0;
+       float r = 0;
 
 	   for (int i = 0; i < this.articulos.size(); i++) {
 		   r += this.articulos.get(i).getPrecio() * this.cantidades.get(i);
